Deletion bug fixed and simple frontend added.

+++ src/main/java/com/cvproject/netsentience/controller/DeviceController.java
@@ -4,8 +4,10 @@ import com.cvproject.netsentience.dto.DeviceUptimeDTO;
 import com.cvproject.netsentience.model.Device;
 import com.cvproject.netsentience.repository.DeviceRepository;
 import com.cvproject.netsentience.repository.MonitoringLogRepository;
+import com.cvproject.netsentience.repository.NatRuleRepository;
 import com.cvproject.netsentience.service.NetworkMonitorService;
 import jakarta.validation.Valid;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,10 +18,15 @@ public class DeviceController {
 
     private final DeviceRepository deviceRepository;
     private final NetworkMonitorService networkService;
+    private final MonitoringLogRepository logRepository;
+    private final NatRuleRepository natRepository;
 
-    public DeviceController(DeviceRepository deviceRepository, NetworkMonitorService networkService) {
+    public DeviceController(DeviceRepository deviceRepository, NetworkMonitorService networkService,
+                            MonitoringLogRepository logRepository, NatRuleRepository natRepository) {
         this.deviceRepository = deviceRepository;
         this.networkService = networkService;
+        this.logRepository = logRepository;
+        this.natRepository = natRepository;
     }
 
     @GetMapping("/{id}/uptime")
@@ -41,8 +48,19 @@ public class DeviceController {
 
     // DELETE /api/devices/{id}
     @DeleteMapping("/{id}")
+    @Transactional
     public void deleteDevice(@PathVariable Long id) {
+        if (!deviceRepository.existsById(id)) {
+            throw new RuntimeException("Device not found");
+        }
+
+        logRepository.deleteByDeviceId(id);
+
+        natRepository.deleteByDeviceId(id);
+
         deviceRepository.deleteById(id);
+
+        System.out.println("Deleted device and history for ID: " + id);
     }
